using worker-loader plugin with inline option without fallback, fix for #1
