+++ package.json
@@ -37,8 +37,8 @@
     "@babel/core": "^7.9.0",
     "@babel/plugin-transform-runtime": "^7.9.0",
     "@babel/preset-env": "^7.9.5",
-    "babel-loader": "^8.1.0",
     "axios": "^0.19.2",
+    "babel-loader": "^8.1.0",
     "eslint": "^6.8.0",
     "eslint-config-standard": "^14.1.1",
     "eslint-plugin-import": "^2.20.2",
@@ -52,11 +52,12 @@
     "stats.js": "^0.17.0",
     "three": "^0.115.0",
     "webpack": "^4.42.1",
-    "webpack-cli": "^3.3.11"
+    "webpack-cli": "^3.3.11",
+    "worker-loader": "^2.0.0"
   },
   "scripts": {
     "dev-es6": "webpack --mode development --progress --colors --watch",
-    "build-es6": "webpack --mode production --optimize-minimize -d",
+    "build-es6": "webpack --mode production --optimize-minimize",
     "build-jshint": "grunt jshint --force",
     "watch": "./node_modules/.bin/watch 'npm run build' ./js/"
   },
