+++ webpack.config.js
@@ -41,6 +41,13 @@ module.exports = {
             ]
           }
         }]
+      },
+      {
+        test: /\worker\.js$/,
+        use: {
+          loader: 'worker-loader',
+          options: { inline: true, fallback: false }
+        }
       }
     ]
   },
